Fix after merge from dev
