+++ team_retarded/src/main/java/application/core/responses/CoreResponse.java
@@ -1,4 +1,4 @@
-package application.core.responses.product;
+package application.core.responses;
 
 import java.util.List;
