+++ team_retarded/src/main/java/application/core/responses/category/AddCategoryResponse.java
@@ -1,6 +1,6 @@
 package application.core.responses.category;
 
-import application.core.responses.product.CoreResponse;
+import application.core.responses.CoreResponse;
 
 public class AddCategoryResponse extends CoreResponse {
 }
