+++ team_retarded/src/main/java/application/core/responses/product/FilterProductsByNameResponse.java
@@ -1,5 +1,7 @@
 package application.core.responses.product;
 
+import application.core.responses.CoreError;
+import application.core.responses.CoreResponse;
 import application.items.Product;
 
 import java.util.List;
@@ -8,7 +10,7 @@ public class FilterProductsByNameResponse extends CoreResponse {
 
     private final List<Product> filterProductsByName;
 
-    public FilterProductsByNameResponse(List<CoreError> errors,List <Product> filterProductsByName) {
+    public FilterProductsByNameResponse(List<CoreError> errors, List<Product> filterProductsByName) {
         super(errors);
         this.filterProductsByName = filterProductsByName;
     }
