+++ team_retarded/src/main/java/application/core/responses/product/SearchProductResponse.java
@@ -1,5 +1,7 @@
 package application.core.responses.product;
 
+import application.core.responses.CoreError;
+import application.core.responses.CoreResponse;
 import application.items.Product;
 
 import java.util.List;
