+++ team_retarded/src/main/java/application/core/responses/product/ShowAllProductsResponse.java
@@ -1,5 +1,6 @@
 package application.core.responses.product;
 
+import application.core.responses.CoreResponse;
 import application.items.Product;
 
 import java.util.List;
