+++ team_retarded/src/main/java/application/core/services/validators/product/FilterProductsByNameValidator.java
@@ -1,7 +1,7 @@
 package application.core.services.validators.product;
 
 import application.core.requests.product.FilterProductsByNameRequest;
-import application.core.responses.product.CoreError;
+import application.core.responses.CoreError;
 
 import java.util.ArrayList;
 import java.util.List;
